refactor DonationForm to use handlePriceInput event handler
