+++ src/CreateShowing/event-handlers.ts
@@ -134,16 +134,3 @@ export async function handleEventFormSubmit(e: FormSubmitEvent, film: string, im
 export function handlePriceTypeInput(e: ChangeEvent, setPriceType: StringStateSetter) {
     setPriceType(e.target.value)
 }
-
-export function handlePriceInput(e: ChangeEvent, setPriceInput: StringStateSetter, setError: StringStateSetter) {
-    setError("")
-    const priceInput = e.target.value
-    const nonPriceRegex = /^\d{0,3}(\.\d{0,2})?$/
-    const isInvalidCharacter = nonPriceRegex.test(priceInput)
-
-    if (!isInvalidCharacter) {
-        setError("Invalid price entered (max. £1000)")
-        return
-    }
-    setPriceInput(priceInput)
-}
