+++ src/SingleShowing/DonationForm.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react"
-import { handleDonationInput } from "./event-handlers"
 import { StringStateSetter } from "../../types"
+import { handlePriceInput } from "../event-handler"
 
 function DonationForm({ donation, setDonation }: {donation: string, setDonation: StringStateSetter}) {
     const [error, setError] = useState("")
@@ -8,7 +8,7 @@ function DonationForm({ donation, setDonation }: {donation: string, setDonation:
     return <>
         <form>
             <label htmlFor="price">Donation (£)</label>
-            <input id="price" type="number" value={donation} placeholder="0" onChange={e => handleDonationInput(e, setDonation, setError)}/>
+            <input id="price" type="number" value={donation} placeholder="0" onChange={e => handlePriceInput(e, setDonation, setError)}/>
         </form>
         {error ? <p>{error}</p> : null}
     </>
