+++ src/event-handler.ts
@@ -4,3 +4,16 @@ import { StringStateSetter } from "../types";
 export function handleTextInput(e: ChangeEvent, setTextInput: StringStateSetter) {
     setTextInput(e.target.value);
 }
+
+export function handlePriceInput(e: ChangeEvent, setPriceInput: StringStateSetter, setError: StringStateSetter) {
+    setError("")
+    const priceInput = e.target.value
+    const nonPriceRegex = /^\d{0,3}(\.\d{0,2})?$/
+    const isInvalidCharacter = nonPriceRegex.test(priceInput)
+
+    if (!isInvalidCharacter) {
+        setError("Invalid price entered (max. £1000)")
+        return
+    }
+    setPriceInput(priceInput)
+}
